Moved modal to separate component
